Function proc_macro_util::add_bounds_to_generics

pub fn add_bounds_to_generics(
    generics: &mut Generics,
    bounds_to_add: impl IntoIterator<Item = TypeParamBound> + Debug,
    generic_param_ident: Option<&Ident>
)

Adds all trait bounds to all generic params. Skips adding bound to param if it already exists. A bound is considered to already exist if:

for lifetimes, exact equality between bounds is found

for types, either the existing bound or the bound to add is a matched suffix of the other

Example

use proc_macro_util::add_bounds_to_generics;
use quote::quote;
use syn::parse2;

// check that a matched path will of a bound on a param will prevent it from being added to param
let mut generics = parse2(quote!(<A: Send, B: std::fmt::Debug>)).unwrap();
let bounds_to_add = [parse2(quote!(Debug)).unwrap()];

add_bounds_to_generics(&mut generics, bounds_to_add, None);

assert_eq!(parse2::<syn::Generics>(quote!(<A: Send + Debug, B: std::fmt::Debug>)).unwrap(), generics);

// check that a matched path will of a bound on a param will prevent it from being added to param
let mut generics = parse2(quote!(<A: Send, B: Debug>)).unwrap();
let bounds_to_add = [parse2(quote!(std::fmt::Debug)).unwrap()];

add_bounds_to_generics(&mut generics, bounds_to_add, None);

assert_eq!(parse2::<syn::Generics>(quote!(<A: Send + std::fmt::Debug, B: Debug>)).unwrap(), generics);

// check that a bound will be added to all params when none of them have it already
let mut generics = parse2(quote!(<A: Send, B: Debug>)).unwrap();
let bounds_to_add = [parse2(quote!(Sync)).unwrap()];

add_bounds_to_generics(&mut generics, bounds_to_add, None);

assert_eq!(parse2::<syn::Generics>(quote!(<A: Send + Sync, B: Debug + Sync>)).unwrap(), generics);

// check that specific idents can have bounds added by providing their ident
let mut generics = parse2(quote!(<A: Send, B: Debug>)).unwrap();
let bounds_to_add = [parse2(quote!(Sync)).unwrap()];
let ident: syn::Ident = parse2(quote!(A)).unwrap();

add_bounds_to_generics(&mut generics, bounds_to_add, Some(&ident));

assert_eq!(parse2::<syn::Generics>(quote!(<A: Send + Sync, B: Debug>)).unwrap(), generics);